env/find_one.py: Change task to find any object of a type
Expert trajectory provided is to the nearest object of that type. Target
object types are chosen with all types of objects in the scene equally
weighted, instead of sampling weighted by the number of objects of that
type in the scene (i.e. if there are more instances of an object in a
scene then it is NOT more likely to choose an object of that type).